Improved Buttons with a SetParts() function, which takes three textures - a corner, edge and center - to create a list of sprites to draw, stretching the edges and center to fit based on the sprites' frameWidth and Height.
Added icons to buttons, using RPG Maker's IconSet sprite sheet for images

Changed the formatting slightly on where Buttons appear on screen, with the pointer to the left of them, instead of inside the Button's texture

Also added a Box class, which the Button class now inherits from. And will be used for creating housings for the Buttons which can't be pressed, and don't have icons or actions assigned to them.